@@ -11,16 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.10",
+    "@angular/animations": "18.2.13",
     "@angular/cdk": "17.3.10",
-    "@angular/common": "17.3.10",
-    "@angular/compiler": "17.3.10",
-    "@angular/core": "17.3.10",
-    "@angular/forms": "17.3.10",
+    "@angular/common": "18.2.13",
+    "@angular/compiler": "18.2.13",
+    "@angular/core": "18.2.13",
+    "@angular/forms": "18.2.13",
     "@angular/material": "17.3.10",
-    "@angular/platform-browser": "17.3.10",
-    "@angular/platform-browser-dynamic": "17.3.10",
-    "@angular/router": "17.3.10",
+    "@angular/platform-browser": "18.2.13",
+    "@angular/platform-browser-dynamic": "18.2.13",
+    "@angular/router": "18.2.13",
     "@fontsource/raleway": "^5.0.17",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
     "@ngneat/until-destroy": "^10.0.0",
@@ -40,10 +40,10 @@
     "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.3",
-    "@angular/cli": "~17.3.3",
-    "@angular/compiler-cli": "17.3.10",
-    "@angular/localize": "17.3.10",
+    "@angular-devkit/build-angular": "^18.2.14",
+    "@angular/cli": "~18.2.14",
+    "@angular/compiler-cli": "18.2.13",
+    "@angular/localize": "18.2.13",
     "@ngxs/devtools-plugin": "^3.8.2",
     "@types/jasmine": "5.1.5",
     "@types/node": "^20.12.3",
