@@ -1,5 +1,5 @@
 import { CommonModule } from "@angular/common";
-import { HttpClientModule } from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import { NgModule } from "@angular/core";
 import { ReactiveFormsModule } from "@angular/forms";
 import { SharedUiModule } from "src/shared-ui/shared-ui.module";
@@ -11,18 +11,12 @@ import { TagFormComponent } from "./tag-form/tag-form.component";
 import { TagTableComponent } from "./tag-table/tag-table.component";
 import { TagsRoutingModule } from "./tags-routing.module";
 
-@NgModule({
-  declarations: [TagTableComponent, TagFormComponent],
-  imports: [
-    CommonModule,
-    DirectivesModule,
-    HttpClientModule,
-    InputModule,
-    PipesModule,
-    ReactiveFormsModule,
-    SharedUiModule,
-    TableModule,
-    TagsRoutingModule,
-  ],
-})
+@NgModule({ declarations: [TagTableComponent, TagFormComponent], imports: [CommonModule,
+        DirectivesModule,
+        InputModule,
+        PipesModule,
+        ReactiveFormsModule,
+        SharedUiModule,
+        TableModule,
+        TagsRoutingModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class TagsModule {}
