@@ -1,9 +1,10 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { NgxsModule } from "@ngxs/store";
 import { AboutState } from "../../store/about.state";
 
 import { AboutComponent } from "./about.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("AboutComponent", () => {
   let component: AboutComponent;
@@ -11,8 +12,9 @@ describe("AboutComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [AboutComponent, HttpClientTestingModule, NgxsModule.forRoot([AboutState])],
-    })
+    imports: [AboutComponent, NgxsModule.forRoot([AboutState])],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+})
       .compileComponents();
 
     fixture = TestBed.createComponent(AboutComponent);
