@@ -1,21 +1,20 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
 import { NgxsModule } from "@ngxs/store";
 import { ApiModule } from "../open-api";
 import { AppComponent } from "./app.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("AppComponent", () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule,
+    declarations: [AppComponent],
+    imports: [RouterTestingModule,
         NgxsModule.forRoot([]),
-        HttpClientTestingModule,
-        ApiModule,
-      ],
-      declarations: [AppComponent],
-    }).compileComponents();
+        ApiModule],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+}).compileComponents();
   });
 
   it("should create the app", () => {
