@@ -1,5 +1,5 @@
 import { CommonModule } from "@angular/common";
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { MatCardModule } from "@angular/material/card";
 import { MatDialogModule } from "@angular/material/dialog";
@@ -14,6 +14,7 @@ import { SummaryCardComponent } from "../../shared-ui/summary-card/summary-card.
 import { GroupState } from "../../store";
 import { DashboardRoutingModule } from "../dashboard-routing.module";
 import { DashboardComponent } from "./dashboard.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("DashboardComponent", () => {
   let component: DashboardComponent;
@@ -37,27 +38,26 @@ describe("DashboardComponent", () => {
       } as Dashboard,
     ];
     await TestBed.configureTestingModule({
-      declarations: [DashboardComponent, SummaryCardComponent],
-      imports: [
-        ApiModule,
+    declarations: [DashboardComponent, SummaryCardComponent],
+    imports: [ApiModule,
         CommonModule,
         DashboardRoutingModule,
-        HttpClientTestingModule,
         MatCardModule,
         MatDialogModule,
         MatListModule,
         NgxsModule.forRoot([GroupState, DashboardState]),
-        PipesModule,
-      ],
-      providers: [
+        PipesModule],
+    providers: [
         {
-          provide: ActivatedRoute,
-          useValue: {
-            params: new BehaviorSubject<Params>({ id: "1" }),
-          },
+            provide: ActivatedRoute,
+            useValue: {
+                params: new BehaviorSubject<Params>({ id: "1" }),
+            },
         },
-      ],
-    }).compileComponents();
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting(),
+    ]
+}).compileComponents();
 
     store = TestBed.inject(Store);
 
