@@ -1,5 +1,5 @@
 import { CommonModule } from "@angular/common";
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { ReactiveFormsModule } from "@angular/forms";
@@ -7,6 +7,7 @@ import { ActivatedRoute } from "@angular/router";
 import { NgxsModule } from "@ngxs/store";
 
 import { GroupDetailsComponent } from "./group-details.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("GroupDetailsComponent", () => {
   let component: GroupDetailsComponent;
@@ -14,23 +15,24 @@ describe("GroupDetailsComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [GroupDetailsComponent],
-      imports: [ReactiveFormsModule, HttpClientTestingModule, CommonModule, NgxsModule.forRoot([])],
-      providers: [
+    declarations: [GroupDetailsComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    imports: [ReactiveFormsModule, CommonModule, NgxsModule.forRoot([])],
+    providers: [
         {
-          provide: ActivatedRoute,
-          useValue: {
-            snapshot: {
-              data: {
-                group: { id: 1 },
-              },
+            provide: ActivatedRoute,
+            useValue: {
+                snapshot: {
+                    data: {
+                        group: { id: 1 },
+                    },
+                },
             },
-          },
         },
-
-      ],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-    })
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting(),
+    ]
+})
       .compileComponents();
 
     fixture = TestBed.createComponent(GroupDetailsComponent);
