@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { ReactiveFormsModule } from "@angular/forms";
@@ -8,6 +8,7 @@ import { ImportType } from "../../open-api";
 import { PipesModule } from "../../pipes";
 
 import { ImportFormComponent } from "./import-form.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("ImportFormComponent", () => {
   let component: ImportFormComponent;
@@ -15,19 +16,18 @@ describe("ImportFormComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ImportFormComponent],
-      imports: [
-        HttpClientTestingModule,
-        ReactiveFormsModule,
+    declarations: [ImportFormComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    imports: [ReactiveFormsModule,
         MatDialogModule,
         PipesModule,
-        NgxsModule.forRoot([])
-      ],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-      providers: [
-        { provide: MatDialogRef, useValue: {} }
-      ]
-    })
+        NgxsModule.forRoot([])],
+    providers: [
+        { provide: MatDialogRef, useValue: {} },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+})
       .compileComponents();
 
     fixture = TestBed.createComponent(ImportFormComponent);
