@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { MatDialogModule } from "@angular/material/dialog";
 import { MatSnackBarModule } from "@angular/material/snack-bar";
@@ -6,6 +6,7 @@ import { NgxsModule, Store } from "@ngxs/store";
 import { ToggleIsSidebarOpen } from "src/store/layout.state.actions";
 import { ApiModule } from "../../open-api";
 import { HeaderComponent } from "./header.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("HeaderComponent", () => {
   let component: HeaderComponent;
@@ -13,15 +14,13 @@ describe("HeaderComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [HeaderComponent],
-      imports: [
-        ApiModule,
-        HttpClientTestingModule,
+    declarations: [HeaderComponent],
+    imports: [ApiModule,
         MatDialogModule,
         MatSnackBarModule,
-        NgxsModule.forRoot([]),
-      ],
-    }).compileComponents();
+        NgxsModule.forRoot([])],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+}).compileComponents();
 
     fixture = TestBed.createComponent(HeaderComponent);
     component = fixture.componentInstance;
