@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { NoopAnimationsModule } from "@angular/platform-browser/animations";
@@ -8,6 +8,7 @@ import { PromptTableState } from "../../store/prompt-table.state";
 import { TableModule } from "../../table/table.module";
 
 import { PromptTableComponent } from "./prompt-table.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("PromptTableComponent", () => {
   let component: PromptTableComponent;
@@ -15,20 +16,20 @@ describe("PromptTableComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [PromptTableComponent],
-      imports: [TableModule, HttpClientTestingModule, NgxsModule.forRoot([PromptTableState]), NoopAnimationsModule],
-      providers: [{
-        provide: ActivatedRoute,
-        useValue: {
-          snapshot: {
-            data: {
-              prompts: []
+    declarations: [PromptTableComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    imports: [TableModule, NgxsModule.forRoot([PromptTableState]), NoopAnimationsModule],
+    providers: [{
+            provide: ActivatedRoute,
+            useValue: {
+                snapshot: {
+                    data: {
+                        prompts: []
+                    }
+                }
             }
-          }
-        }
-      }],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-    })
+        }, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+})
       .compileComponents();
 
     fixture = TestBed.createComponent(PromptTableComponent);
