@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { ActivatedRoute } from "@angular/router";
 import { NgxsModule } from "@ngxs/store";
@@ -6,6 +6,7 @@ import { SharedUiModule } from "../../shared-ui/shared-ui.module";
 import { TableModule } from "../../table/table.module";
 
 import { ReceiptProcessingSettingsTableComponent } from "./receipt-processing-settings-table.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("ReceiptProcessingSettingsTableComponent", () => {
   let component: ReceiptProcessingSettingsTableComponent;
@@ -13,15 +14,17 @@ describe("ReceiptProcessingSettingsTableComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ReceiptProcessingSettingsTableComponent],
-      imports: [NgxsModule.forRoot(), HttpClientTestingModule, TableModule, SharedUiModule],
-      providers: [
+    declarations: [ReceiptProcessingSettingsTableComponent],
+    imports: [NgxsModule.forRoot(), TableModule, SharedUiModule],
+    providers: [
         {
-          provide: ActivatedRoute,
-          useValue: {}
-        }
-      ]
-    })
+            provide: ActivatedRoute,
+            useValue: {}
+        },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+})
       .compileComponents();
 
     fixture = TestBed.createComponent(ReceiptProcessingSettingsTableComponent);
