@@ -2,11 +2,12 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ItemListComponent } from './item-list.component';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { ActivatedRoute } from '@angular/router';
 import { PipesModule } from 'src/pipes/pipes.module';
 import { FormGroup } from '@angular/forms';
 import { NgxsModule } from '@ngxs/store';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('ItemListComponent', () => {
   let component: ItemListComponent;
@@ -14,16 +15,18 @@ describe('ItemListComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ItemListComponent],
-      imports: [HttpClientTestingModule, PipesModule, NgxsModule.forRoot([])],
-      providers: [
+    declarations: [ItemListComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    imports: [PipesModule, NgxsModule.forRoot([])],
+    providers: [
         {
-          provide: ActivatedRoute,
-          useValue: { snapshot: { data: {} } },
+            provide: ActivatedRoute,
+            useValue: { snapshot: { data: {} } },
         },
-      ],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-    }).compileComponents();
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting(),
+    ]
+}).compileComponents();
 
     fixture = TestBed.createComponent(ItemListComponent);
     component = fixture.componentInstance;
