@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { ReactiveFormsModule } from "@angular/forms";
@@ -16,6 +16,7 @@ import { SnackbarService } from "../../services";
 import { QueueMode } from "../../services/receipt-queue.service";
 import { GroupState } from "../../store";
 import { ReceiptFormComponent } from "./receipt-form.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("ReceiptFormComponent", () => {
   let component: ReceiptFormComponent;
@@ -23,27 +24,26 @@ describe("ReceiptFormComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ReceiptFormComponent],
-      imports: [
-        ApiModule,
+    declarations: [ReceiptFormComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    imports: [ApiModule,
         PipesModule,
-        HttpClientTestingModule,
         MatDialogModule,
         MatSnackBarModule,
         NgxsModule.forRoot([GroupState]),
         NoopAnimationsModule,
         PipesModule,
         ReactiveFormsModule,
-        SharedUiModule,
-      ],
-      providers: [
+        SharedUiModule],
+    providers: [
         {
-          provide: ActivatedRoute,
-          useValue: { snapshot: { data: {}, queryParams: {} }, params: of({}) },
+            provide: ActivatedRoute,
+            useValue: { snapshot: { data: {}, queryParams: {} }, params: of({}) },
         },
-      ],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-    }).compileComponents();
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting(),
+    ]
+}).compileComponents();
 
     fixture = TestBed.createComponent(ReceiptFormComponent);
     component = fixture.componentInstance;
