@@ -1,10 +1,11 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { MatSnackBarModule } from "@angular/material/snack-bar";
 import { ActivatedRoute } from "@angular/router";
 import { ApiModule } from "../../open-api";
 import { UploadImageComponent } from "./upload-image.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("UploadImageComponent", () => {
   let component: UploadImageComponent;
@@ -12,13 +13,15 @@ describe("UploadImageComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [UploadImageComponent],
-      imports: [ApiModule, MatSnackBarModule, HttpClientTestingModule],
-      providers: [
+    declarations: [UploadImageComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    imports: [ApiModule, MatSnackBarModule],
+    providers: [
         { provide: ActivatedRoute, useValue: { snapshot: { data: {} } } },
-      ],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-    }).compileComponents();
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting(),
+    ]
+}).compileComponents();
 
     fixture = TestBed.createComponent(UploadImageComponent);
     component = fixture.componentInstance;
