@@ -1,8 +1,9 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
 import { ResolveFn } from "@angular/router";
 import { ApiModule, Category, CategoryService } from "../open-api";
 import { categoryResolverFn } from "./categories.resolver";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("CategoriesResolverService", () => {
   const executeResolver: ResolveFn<Category[]> = (
@@ -14,8 +15,9 @@ describe("CategoriesResolverService", () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [ApiModule, HttpClientTestingModule],
-    });
+    imports: [ApiModule],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
   });
 
   it("should call get all categories", () => {
