@@ -1,8 +1,9 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
 import { ResolveFn } from "@angular/router";
 import { ApiModule, Tag, TagService } from "../open-api";
 import { tagResolverFn } from "./tags.resolver";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("TagsResolverService", () => {
   const executeResolver: ResolveFn<Tag[]> = (
@@ -14,8 +15,9 @@ describe("TagsResolverService", () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [ApiModule, HttpClientTestingModule],
-    });
+    imports: [ApiModule],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
   });
 
   it("should call tag service", () => {
