@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { ReactiveFormsModule } from "@angular/forms";
@@ -8,6 +8,7 @@ import { NgxsModule } from "@ngxs/store";
 import { PipesModule } from "src/pipes/pipes.module";
 import { ApiModule } from "../../open-api";
 import { ResetPasswordComponent } from "./reset-password.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("ResetPasswordComponent", () => {
   let component: ResetPasswordComponent;
@@ -15,20 +16,17 @@ describe("ResetPasswordComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ResetPasswordComponent],
-      imports: [
-        ApiModule,
-        HttpClientTestingModule,
+    declarations: [ResetPasswordComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    imports: [ApiModule,
         PipesModule,
         MatDialogModule,
         MatSnackBarModule,
         NgxsModule.forRoot([]),
         PipesModule,
-        ReactiveFormsModule,
-      ],
-      providers: [{ provide: MatDialogRef, useValue: {} }],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-    }).compileComponents();
+        ReactiveFormsModule],
+    providers: [{ provide: MatDialogRef, useValue: {} }, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+}).compileComponents();
 
     fixture = TestBed.createComponent(ResetPasswordComponent);
     component = fixture.componentInstance;
