@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { MatDialogModule } from "@angular/material/dialog";
@@ -7,6 +7,7 @@ import { NgxsModule } from "@ngxs/store";
 import { TableModule } from "src/table/table.module";
 import { ApiModule } from "../../open-api";
 import { UserListComponent } from "./user-list.component";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("UserListComponent", () => {
   let component: UserListComponent;
@@ -14,17 +15,15 @@ describe("UserListComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [UserListComponent],
-      imports: [
-        ApiModule,
-        HttpClientTestingModule,
+    declarations: [UserListComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    imports: [ApiModule,
         MatDialogModule,
         MatSnackBarModule,
         NgxsModule.forRoot([]),
-        TableModule,
-      ],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-    }).compileComponents();
+        TableModule],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+}).compileComponents();
 
     fixture = TestBed.createComponent(UserListComponent);
     component = fixture.componentInstance;
